+++ R/peaks.R
@@ -26,13 +26,16 @@ NULL
 #' p-value, q-value, and fold-change information for each peak will be lost if
 #' combining peaks.
 #' @param broad Call broad peaks (\code{--broad} parameter for MACS)
+#' @param format File format to use. Should be either "BED" or "BEDPE" (see 
+#' MACS documentation).
 #' @param outdir Path for output files
 #' @param fragment.tempdir Path to write temporary fragment files. Only used if
 #' \code{group.by} is not NULL.
 #' @param effective.genome.size Effective genome size parameter for MACS
 #' (\code{-g}). Default is the human effective genome size (2.7e9).
-#' @param extsize \code{extsize} parameter for MACS.
-#' @param shift \code{shift} parameter for MACS.
+#' @param extsize \code{extsize} parameter for MACS. Only relevant if 
+#' format="BED"
+#' @param shift \code{shift} parameter for MACS. Only relevant if format="BED"
 #' @param additional.args Additional arguments passed to MACS. This should be a
 #' single character string
 #' @param name Name for output MACS files. This will also be placed in the
@@ -58,6 +61,7 @@ CallPeaks.Seurat <- function(
   idents = NULL,
   macs2.path = NULL,
   broad = FALSE,
+  format = "BED",
   outdir = tempdir(),
   fragment.tempdir = tempdir(),
   combine.peaks = TRUE,
@@ -124,6 +128,7 @@ CallPeaks.Seurat <- function(
         macs2.path = macs2.path,
         outdir = outdir,
         broad = broad,
+        format = format,
         effective.genome.size = effective.genome.size,
         extsize = extsize,
         shift = shift,
@@ -164,6 +169,7 @@ CallPeaks.Seurat <- function(
       macs2.path = macs2.path,
       outdir = outdir,
       broad = broad,
+      format = format,
       effective.genome.size = effective.genome.size,
       extsize = extsize,
       shift = shift,
@@ -186,6 +192,7 @@ CallPeaks.ChromatinAssay <- function(
   macs2.path = NULL,
   outdir = tempdir(),
   broad = FALSE,
+  format = "BED",
   effective.genome.size = 2.7e9,
   extsize = 200,
   shift = -extsize/2,
@@ -204,6 +211,7 @@ CallPeaks.ChromatinAssay <- function(
     macs2.path = macs2.path,
     outdir = outdir,
     broad = broad,
+    format = format,
     effective.genome.size = effective.genome.size,
     extsize = extsize,
     shift = shift,
@@ -225,6 +233,7 @@ CallPeaks.Fragment <- function(
   macs2.path = NULL,
   outdir = tempdir(),
   broad = FALSE,
+  format = "BED",
   effective.genome.size = 2.7e9,
   extsize = 200,
   shift = -extsize/2,
@@ -240,6 +249,7 @@ CallPeaks.Fragment <- function(
     macs2.path = macs2.path,
     outdir = outdir,
     broad = broad,
+    format = format,
     effective.genome.size = effective.genome.size,
     extsize = extsize,
     shift = shift,
@@ -263,6 +273,7 @@ CallPeaks.default <- function(
   macs2.path = NULL,
   outdir = tempdir(),
   broad = FALSE,
+  format = "BED",
   effective.genome.size = 2.7e9,
   extsize = 200,
   shift = -extsize/2,
@@ -296,7 +307,16 @@ CallPeaks.default <- function(
   }
 
   broadstring <- ifelse(test = broad, yes = " --broad ", no = "")
-
+  nomod_str <- ifelse(
+    test = format == "BED",
+    yes = paste0(" --nomodel --extsize ",
+    as.character(x = extsize),
+    " --shift ",
+    as.character(x = shift)
+    ),
+    no = ""
+  )
+  
   cmd <- paste0(
     macs2.path,
     " callpeak -t ",
@@ -304,10 +324,9 @@ CallPeaks.default <- function(
     " -g ",
     as.character(x = effective.genome.size),
     broadstring,
-    " -f BED --nomodel --extsize ",
-    as.character(x = extsize),
-    " --shift ",
-    as.character(x = shift),
+    " -f ",
+    format,
+    nomod_str,
     " -n ",
     "'",
     as.character(x = name),